config-loader: move transform logic into separate lib folder
