+++ packages/config-loader/src/lib/transform/index.ts
@@ -14,6 +14,5 @@
  * limitations under the License.
  */
 
-export { readEnvConfig } from './env';
-export * from './transform';
-export * from './schema';
+export { applyConfigTransforms } from './apply';
+export { createIncludeTransform } from './include';
